requirements.txt
requests
python-dotenv
pandas
numpy
yfinance
anthropic
schedule
python-telegram-bot
finvizfinance
